Move welcome related stuff to WelcomeModule.
