+++ src/main/webapp/resources/ng2/app/app.module.ts
@@ -1,23 +1,25 @@
 import {NgModule} from "@angular/core";
 import {BrowserModule} from "@angular/platform-browser";
-import {NgbModule} from "@ng-bootstrap/ng-bootstrap";
-import {AppComponent} from "./app.component";
-import {WelcomeComponent} from "./welcome.component";
-import {NavComponent} from "./nav.component";
-import {FooterComponent} from "./footer.component";
-import {routing, appRoutingProviders} from "./app.routing";
-import {PetListComponent} from "./pet-list.component";
 import {FormsModule} from "@angular/forms";
 import {HttpModule} from "@angular/http";
+import {NgbModule} from "@ng-bootstrap/ng-bootstrap";
+import {AppComponent} from "./app.component";
+import {NavComponent} from "./nav.component";
+import {FooterComponent} from "./footer.component";
+import {PetListComponent} from "./pets/pet-list.component";
+import {routing, appRoutingProviders} from "./app.routing";
 import {OwnersModule} from "./owners/owners.module";
+import {WelcomeModule} from "./welcome/welcome.module";
 
 @NgModule({
-  imports: [BrowserModule, FormsModule, HttpModule, NgbModule, routing, OwnersModule],
+  imports: [
+    BrowserModule, FormsModule, HttpModule, NgbModule,
+    routing, OwnersModule, WelcomeModule
+  ],
   providers: [appRoutingProviders],
   bootstrap: [AppComponent],
   declarations: [
     AppComponent,
-    WelcomeComponent,
     NavComponent,
     FooterComponent,
     PetListComponent
