+++ src/main/webapp/resources/ng2/app/app.routing.ts
@@ -1,13 +1,9 @@
 import {Routes, RouterModule} from "@angular/router";
-import {WelcomeComponent} from "./welcome.component";
 import {ModuleWithProviders} from "@angular/core";
-import {PetListComponent} from "./pet-list.component";
-import {OwnersComponent} from "./owners/owners.component";
+import {PetListComponent} from "./pets/pet-list.component";
 
 const appRoutes: Routes = [
   {path: '', pathMatch: 'full', redirectTo: '/welcome'},
-  {path: 'welcome', component: WelcomeComponent},
-  {path: 'owners', component: OwnersComponent},
   {path: 'pets', component: PetListComponent},
 ];
