+++ src/main/webapp/resources/ng2/app/welcome/welcome.component.html
