+++ src/main/webapp/resources/ng2/app/welcome/welcome.routing.ts
@@ -1,0 +1,9 @@
+import {Routes, RouterModule} from "@angular/router";
+import {ModuleWithProviders} from "@angular/core";
+import {WelcomeComponent} from "./welcome.component";
+
+const appRoutes: Routes = [
+  {path: 'welcome', component: WelcomeComponent},
+];
+
+export const welcomeRouting: ModuleWithProviders = RouterModule.forChild(appRoutes);
